Time: 11 ms (29.74%), Space: 14.2 MB (78.83%) - LeetHub

+++ 1482-how-many-numbers-are-smaller-than-the-current-number/1482-how-many-numbers-are-smaller-than-the-current-number.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     vector<int> smallerNumbersThanCurrent(vector<int>& nums) {
-        int numsSize = nums.size();
+        short numsSize = nums.size();
         vector<int> answer(numsSize);
 
-        for (int i = 0; i < numsSize - 1; ++i)
+        for (short i = 0; i < numsSize - 1; ++i)
         {
-            for (int j = i + 1; j < numsSize; ++j)
+            for (short j = i + 1; j < numsSize; ++j)
             {
                 if (nums[i] > nums[j])
                     ++answer[i];
